pyproject.toml
[tool.poetry]
name = "gemini-python"
version = "0.1.0"
description = "Gemini is an automatic random testing tool for ScyllaDB."
authors = ["Lukasz Sojka <lukasz.sojka@scylladb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cassandra-driver = "^3.25.0"
coverage = "^6.4.4"


[tool.poetry.group.dev.dependencies]
pylint = "^2.14.5"
black = "^22.6.0"
mypy = "^0.971"
pytest = "^7.1.2"
pre-commit = "^2.20.0"

[tool.poetry.scripts]
gemini = 'gemini_python.console:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.pylint.'MESSAGES CONTROL']
disable = "missing-module-docstring, missing-function-docstring, fixme, too-few-public-methods, too-many-arguments"

[tool.pylint.'FORMAT']
# black formats anyway to 100, but not long strings. For these allow len of 140.
max-line-length=140

[tool.mypy]
python_version = "3.10"
strict = true
disallow_any_generics = false
exclude = ["notes"]

[[tool.mypy.overrides]]
module = 'tests.*'
disallow_untyped_defs = false
